modelo de reviews enviado para o server
